Modified mk-pve.sh shell script file for pgAdmin4 version 4.25 	 - Replaced the "requirements.txt" file using file from v4.25 	 - Moved checking for existing PVE to top of script 	 - Changed some of the comments -- we have too many -> reduce me! 	 - Changed pip3 line: 	   *) removed psycopg2 : now included in requirements.txt file 	   *) specified version 2.2.1 for Sphinx : documents build failed 	      without version above.

+++ pgadmin4-slackbuild/mk-pve.sh
@@ -1,27 +1,92 @@
 #!/bin/sh -e
 
-# This is a simple script to create Python Virtual Environment for pgAdmin4.
+# Python Virtual Environment for pgAdimin4 version 4.25
+# Date: September 13, 2020
+# Author: Wael Hammoudeh - w_hammoudeh -at- hotmail dot com
+
+# This script is to create Python Virtual Environment for pgAdmin4.
+# Python "venv" module is used to create this virtual environment.
 # pgAdmin4 developers recommend running it under Python Virtual Environment.
 # I also build pgAdmin4 slackware package under this Python Virtual Environment.
 # ------------------------------------------------------------------
+
 # NOTE ***** INTERNET CONNECTION IS REQUIRED FOR THIS SCRIPT *******
 #      *****     NO CONNECTION CHECKING IS DONE HERE !!!     ******* 
 # ------------------------------------------------------------------
 
-# Python Virtual Environment ROOT "pve" is created under /usr/local directory.
-# This makes "PVE_ROOT=/usr/local/pve".
-# Script makes application virtual environment directory under PVE_ROOT; it 
-# becomes: "APP_PVE=/usr/local/pve/pgAdmin4-pve", script then installs all
-# required python packages plus "sphinx" - required to make documentation.
-# If script finds APP_PVE directory, it exits and does nothing.
+# Python Virtual Environment PVE_ROOT is created under /usr/local directory:
+# PVE_ROOT=/usr/local/pve
+# Application virtual environment directory under PVE_ROOT:
+# APP_PVE=/usr/local/pve/pgAdmin4-pve
+# Script installs all required python packages plus "sphinx" - required to make
+# documentation.
+# If script finds APP_PVE directory, it exits and does nothing. So if you are
+# upgrading pgAdimin4 to newer version make sure to remove APP_PVE dirctory first.
 #
-# See _create_python_virtualenv() function in SRC/pkg/linux/build-functions.sh
+# This script is based on _create_python_virtualenv() function found in source in
+# SOURCE/pkg/linux/build-functions.sh file.
+
+PVE_ROOT=/usr/local/pve
+APP_PVE=$PVE_ROOT/pgAdmin4-pve
+
+if [ -d $APP_PVE ]; then
+  echo ""
+  echo "Found existing pgAdmin4 python virtual environment at:"
+  echo ""
+  echo "     $APP_PVE"
+  echo ""
+  echo "If you want to update pgAdmin4 python virtual environment,"
+  echo "make sure that you have internet connection and you no longer"
+  echo "need the existing environment; then remove its directory with:"
+  echo " rm -rf $APP_PVE"
+  echo "Afterward run this script again."
+  echo "Exiting."
+  echo ""
+  exit;
+fi
+
+# Is postgresql package installed?
+PG_PKG=$(ls /var/lib/pkgtools/packages/postgresql*)
+
+# empty string from ls command above indicates missing PostgreSQL package
+if [ -z "$PG_PKG" ]; then
+    echo "PostgreSQL package not found. Please install PostgreSQL with"
+    echo "slackpkg. Exiting!"
+    exit 1;
+fi
+
+# get postgresql bin directory path; needed to install psycopg2 python module.
+# pg_config file is usually installed in postgresql bin directory
+PG_BIN_PART1=$(cat $PG_PKG | egrep "*/bin/pg_config")
+
+# empty string means missed up installation!
+if [ -z "$PG_BIN_PART1" ]; then
+    echo "ERROR: Could not get path for PostgreSQL bin directory."
+    echo "Exiting!"
+    exit 1;
+fi
+
+# prepend leading slash to directory
+PG_BIN_PART2=/$PG_BIN_PART1
+
+# This is pg_config file which MUST exist in file system
+if [ ! -f "$PG_BIN_PART2" ]; then
+    echo "ERROR: Could not find pg_config file."
+    echo "Exiting"
+    exit 1;
+fi
+
+# extract pathname using dirname
+PG_BIN_PATH=$(dirname "$PG_BIN_PART2")
+
+# add postgresql bin directory to PATH
+PATH=$PATH:$PG_BIN_PATH
 
 CWD=$(pwd)
 
 rm -f requirements.txt
 
-# File requirements.txt was copied from pgAmin4-4.22 source tar ball
+# File requirements.txt was copied from pgAmin4-4.25 source tar ball
 install --mode=644 /dev/stdin "$CWD/requirements.txt" <<END
 ###############################################################################
 #
@@ -63,69 +128,18 @@ psycopg2>=2.8
 python-dateutil>=2.8.0
 SQLAlchemy>=1.3.13
 Flask-Security-Too>=3.0.0
-sshtunnel>=0.1.4
+bcrypt<=3.1.7
+sshtunnel>=0.1.5
 ldap3>=2.5.1
 END
-
-# postgresql is required to build pgAdmin4 - not part of slackware-current
-# Is postgresql package installed?
-PG_PKG=$(ls /var/lib/pkgtools/packages/postgresql*)
-
-# empty string from ls command above indecates missing PostgreSQL package
-if [ -z "$PG_PKG" ]; then
-    echo "PostgreSQL package not found. Please install PostgreSQL with"
-    echo "slackpkg. Exiting!"
-    exit 1;
-fi
-
-# get postgresql bin directory path; needed to install psycopg2 python module.
-# pg_config file is usually installed in postgresql bin directory
-PG_BIN_PART1=$(cat $PG_PKG | egrep "*/bin/pg_config")
-
-# empty string means missed up installation!
-if [ -z "$PG_BIN_PART1" ]; then
-    echo "ERROR: Could not get path for PostgreSQL bin directory."
-    echo "Exiting!"
-    exit 1;
-fi
-
-# prepend leading slash to directory
-PG_BIN_PART2=/$PG_BIN_PART1
-
-# This is pg_config file which MUST exist in file system
-if [ ! -f "$PG_BIN_PART2" ]; then
-    echo "ERROR: Could not find pg_config file."
-    echo "Exiting"
-    exit 1;
-fi
-
-# extract pathname using dirname
-PG_BIN_PATH=$(dirname "$PG_BIN_PART2")
-
-# add postgresql bin directory to PATH
-PATH=$PATH:$PG_BIN_PATH
-
-PVE_ROOT=/usr/local/pve
-PY_EXEC=/usr/bin/python3
-APP_PVE=$PVE_ROOT/pgAdmin4-pve
-
-if [ -d $APP_PVE ]; then
-  echo ""
-  echo "Found existing pgAdmin4 python virtual environment at:"
-  echo ""
-  echo "     $APP_PVE"
-  echo ""
-  echo "If you want to update pgAdmin4 python virtual environment,"
-  echo "please remove its directory first."
-  echo "Exiting."
-  echo ""
-  exit;
-fi
 
 # tell user what we are doing
 echo "Making Python Virual Environment ..."
 
-# create root directory if it does not exist
+# use python3
+PY_EXEC=/usr/bin/python3
+
+# create PVE_ROOT directory if it does not exist
 if [ ! -d $PVE_ROOT ]; then
   echo "Creating directory: $PVE_ROOT"
   mkdir $PVE_ROOT
@@ -133,38 +147,38 @@ fi
 
 cd $PVE_ROOT
 
-# create PVE for pgAdimin4 
+# create APP_PVE for pgAdimin4 
 $PY_EXEC -m venv $APP_PVE
 
-echo "Created PVE at $APP_PVE"
+echo "Created APP_PVE at $APP_PVE"
 
 # activate virtual environment to install required modules
 source $APP_PVE/bin/activate
 
 # install required python modules
 # pip install --upgrade pip @@@@ very BAD BAD BAD idea DO NOT DO IT !!!
-# File requirements.txt was copied from pgAmin4-4.19 tar ball
+
 echo "Installing required python MODULES for pgAdmin4"
 
-pip3 install --no-cache-dir --no-binary psycopg2 sphinx -r $CWD/requirements.txt
+pip3 install --no-cache-dir Sphinx==2.2.1 -r $CWD/requirements.txt
 
-# Sphinx is needed to build documentation
-# pip3 install sphinx //moved to above line
+# Sphinx is needed to build documentation -- version 2.2.1 is required here
 
-# source: _create_python_virtualenv() function in SRC/pkg/linux/build-functions.sh
+# source: _create_python_virtualenv() function in SOURCE/pkg/linux/build-functions.sh
 # Figure out some paths for use when completing the venv
 # Use "python3" here as we want the venv path
+
 PYMODULES_PATH=$(python3 -c "from distutils.sysconfig import get_python_lib; print(get_python_lib())")
 DIR_PYMODULES_PATH=`dirname ${PYMODULES_PATH}`
 
 # Use /usr/bin/python3 here as we want the system path
 PYSYSLIB_PATH=$(/usr/bin/python3 -c "import sys; \
                 print('%s/lib64/python%d.%.d' % (sys.prefix, sys.version_info.major, sys.version_info.minor))")
-#
-#
+
 # Symlink in the rest of the Python libs. This is required because the runtime
 # will clear PYTHONHOME for safety, which has the side-effect of preventing
 # it from finding modules that are not explicitly included in the venv
+
 cd ${DIR_PYMODULES_PATH}
 
 # Files
@@ -196,9 +210,9 @@ fi
 
 echo ""
 echo " All done; to use virtual environment just source the script like:" 
-echo " source $APP_PVE/bin/activate"
+echo "   source $APP_PVE/bin/activate"
 echo " and to exit or terminate just issue:"
-echo " deactivate"
+echo "   deactivate"
 echo ""
 
 # be nice
